+++ muparser-2.2.5/include/muParserTemplateMagic.h
@@ -2,7 +2,6 @@
 #define MU_PARSER_TEMPLATE_MAGIC_H
 
 #include <cmath>
-#include "muParserError.h"
 
 namespace mu {
 //-----------------------------------------------------------------------------------------------
