+++ muparser-2.2.5/src/muParserBytecode.cpp
@@ -33,7 +33,6 @@
 #include <vector>
 
 #include "muParserDef.h"
-#include "muParserError.h"
 #include "muParserTemplateMagic.h"
 #include "muParserToken.h"
