+++ muparser-2.2.5/src/muParserError.cpp
@@ -22,7 +22,9 @@
   DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
-#include "muParserError.h"
+#include "muParserDef.h"
+
+#include <assert.h>
 
 namespace mu {
 //---------------------------------------------------------------------------
